调式采集触发器modbus和整个转发流程
1、修改variant不兼容float array问题
2、修改json_write不支持int64类型
3、修改modbus和mqtt时间戳获取和转换格式一至问题
4、完善modbus采集数据转换为数值类型的类

+++ core/data_context.h
@@ -26,7 +26,7 @@ public:
     {
         std::string source_device;
         int stream_id;
-        int64_t timestamp_ms;
+        double timestamp_ms;
         uint64_t frame_index; // 帧序号，算法对齐必用
     } header;
